+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 All notable changes to this package will be documented in this file.
 
 ## [Unreleased]
+- Ability to disable custom UnityObject inspector for better compatibility with other libraries, in MyBox Window settings
 - Added: SceneClickHandler support of Physics2d and Plane
+- Added: MyDefinesUtility simplifies handling of PlayerSettings.Set(Get)ScriptingDefineSymbols
 
 ## [1.8.0] - 2023-10-09
 - Now it is possible to toggle some features that may lead to performance issues in editor in MyBox Window